requirements.txt
pycryptodomex>=3.19.0
pyOpenSSL>=23.3.0
cryptography>=41.0.0
click>=8.1.0
colorama>=0.4.6